+++ src/test/java/com/iexec/sms/config/SecretsConfigTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2023 IEXEC BLOCKCHAIN TECH
+ * Copyright 2023-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,6 +21,8 @@ import com.iexec.sms.secret.MeasuredSecretService;
 import com.iexec.sms.secret.compute.TeeTaskComputeSecretRepository;
 import com.iexec.sms.secret.web2.Web2SecretRepository;
 import com.iexec.sms.secret.web3.Web3SecretRepository;
+import com.iexec.sms.tee.challenge.TeeChallengeRepository;
+import com.iexec.sms.utils.EthereumCredentialsRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
@@ -122,6 +124,36 @@ class SecretsConfigTests {
         assertAll(
                 () -> assertThat(secretsType).isEqualTo("compute"),
                 () -> assertThat(metricsPrefix).isEqualTo("iexec.sms.secrets.compute."),
+                () -> verify(metricsService).registerNewMeasuredSecretService(measuredSecretService)
+        );
+    }
+
+    @Test
+    void teeChallengeMeasuredSecretService() {
+        final TeeChallengeRepository repository = mock(TeeChallengeRepository.class);
+        final MeasuredSecretService measuredSecretService = secretsConfig.teeChallengeMeasuredSecretService(repository, STORED_SECRETS_COUNT_PERIOD);
+
+        final String secretsType = ((String) ReflectionTestUtils.getField(measuredSecretService, "secretsType"));
+        final String metricsPrefix = ((String) ReflectionTestUtils.getField(measuredSecretService, "metricsPrefix"));
+
+        assertAll(
+                () -> assertThat(secretsType).isEqualTo("TEE challenges"),
+                () -> assertThat(metricsPrefix).isEqualTo("iexec.sms.secrets.tee_challenges."),
+                () -> verify(metricsService).registerNewMeasuredSecretService(measuredSecretService)
+        );
+    }
+
+    @Test
+    void ethereumCredentialsMeasuredSecretService() {
+        final EthereumCredentialsRepository repository = mock(EthereumCredentialsRepository.class);
+        final MeasuredSecretService measuredSecretService = secretsConfig.ethereumCredentialsMeasuredSecretService(repository, STORED_SECRETS_COUNT_PERIOD);
+
+        final String secretsType = ((String) ReflectionTestUtils.getField(measuredSecretService, "secretsType"));
+        final String metricsPrefix = ((String) ReflectionTestUtils.getField(measuredSecretService, "metricsPrefix"));
+
+        assertAll(
+                () -> assertThat(secretsType).isEqualTo("Ethereum Credentials"),
+                () -> assertThat(metricsPrefix).isEqualTo("iexec.sms.secrets.ethereum_credentials."),
                 () -> verify(metricsService).registerNewMeasuredSecretService(measuredSecretService)
         );
     }